Refactor hsi_event into its own file, to avoid circular references.
- update tlo/method imports
- update test imports
