+++ src/tlo/methods/cardio_metabolic_disorders.py
@@ -26,7 +26,7 @@ from tlo.methods import Metadata
 from tlo.methods import demography as de
 from tlo.methods.causes import Cause
 from tlo.methods.dxmanager import DxTest
-from tlo.methods.healthsystem import HSI_Event
+from tlo.methods.hsi_event import HSI_Event
 from tlo.methods.symptommanager import Symptom
 from tlo.util import random_date
