+++ src/tlo/methods/hsi_generic_first_appts.py
@@ -23,7 +23,7 @@ from tlo.methods.care_of_women_during_pregnancy import (
 )
 from tlo.methods.chronicsyndrome import HSI_ChronicSyndrome_SeeksEmergencyCareAndGetsTreatment
 from tlo.methods.epilepsy import HSI_Epilepsy_Start_Anti_Epileptic
-from tlo.methods.healthsystem import HSI_Event
+from tlo.methods.hsi_event import HSI_Event
 from tlo.methods.hiv import HSI_Hiv_TestAndRefer
 from tlo.methods.labour import HSI_Labour_ReceivesSkilledBirthAttendanceDuringLabour
 from tlo.methods.measles import HSI_Measles_Treatment
