+++ src/tlo/methods/tb.py
@@ -15,7 +15,8 @@ from tlo.lm import LinearModel, LinearModelType, Predictor
 from tlo.methods import Metadata, hiv
 from tlo.methods.causes import Cause
 from tlo.methods.dxmanager import DxTest
-from tlo.methods.healthsystem import HealthSystemChangeParameters, HSI_Event
+from tlo.methods.healthsystem import HealthSystemChangeParameters
+from tlo.methods.hsi_event import HSI_Event
 from tlo.methods.symptommanager import Symptom
 from tlo.util import random_date
